Add nix package GCC and Xtensa toolchain for ESP32
Required to link the ESP32 binary after rust compilation.

The toolchain is downloaded from the Espressif release page, we don't
compile it from source.

I have tried to compile the toolchain from source. I have tried to IFD to
download the dependencies, it is working for URL but not for git repos
as their build tool (ct-ng) is not able to download the git dependencies
witin the nix sandbox.

 Untracked files: